+++ ui/Gruntfile.js
@@ -10,13 +10,18 @@ module.exports = function(grunt) {
         jitGrunt: {
             protractor: 'grunt-protractor-runner',
             ngAnnotate: 'grunt-ng-annotate',
-            ngtemplates: 'grunt-angular-templates'
+            ngtemplates: 'grunt-angular-templates',
+            'validate-package': 'grunt-nsp-package'
+
         }
     });
 
-    // server dev environment
+    // server dev environment with browsersync
+    grunt.registerTask('autosync', ['ngtemplates:dev', 'concat:dev', 'ngAnnotate', 'uglify:dev', 'wiredep', 'concurrent:auto']);
+    
+    // server dev environment no auto refresh
     grunt.registerTask('serve', ['ngtemplates:dev', 'concat:dev', 'ngAnnotate', 'uglify:dev', 'wiredep', 'concurrent:dev']);
-   
+      
     // runs server side specs and UI specs
     grunt.registerTask('spec', ['jshint', 'jasmine_node', 'karma:unit']);
 
@@ -28,4 +33,7 @@ module.exports = function(grunt) {
 
     // verfies javascript using jshint
     grunt.registerTask('jshinting', ['jshint']);
+
+    // verifies security
+    grunt.registerTask('auditpkg', ['validate-package']);
 };