+++ ui/grunt/wiredep.js
@@ -1,6 +1,6 @@
 
 module.exports = {
       target: {
-        src: 'views/index.html' // point to your HTML file.
+        src: 'views/**/*.html' // point to your HTML file.
       }
 };