+++ ui/lib/controllers/main/pages.js
@@ -2,9 +2,23 @@ var Hapi = require('hapi');
 var debug = require('debug')('frontend:pages');
 
 exports.index = function(request, reply) {
-  reply.view('index', { title: 'Test' });
+
+  request.server.render('partials/header', {}, function(err, rendered, config) {
+    reply.view('index', {
+      title: 'Hejsan Hapi Angular User',
+      navbar: rendered
+    });
+  });
+
 };
 
 exports.profile = function(request, reply) {
-  reply.view('profile', { title: 'Welcome ' + request.auth.credentials.name });
+
+  request.server.render('partials/header', {}, function(err, rendered, config) {
+    reply.view('profile', {
+      title: 'Hejsan Hapi Angular Secure User',
+      navbar: rendered
+    });
+  });
+
 };